requirements_test.txt
aiogithubapi>=24.6.0
aioresponses>=0.7.0
pylint-strict-informational>=0.1
pytest>=8.0.2
pytest-aiohttp>=1.0.0
pytest-homeassistant-custom-component==0.13.307
